InteriorAI/requirements.txt
torch
diffusers
transformers
accelerate
safetensors
opencv-python
numpy
Pillow
